refactor(webui-host): PumpDetailDto/MinerDetailDto組み立てを専用ビルダーへ分割
BlockDetailDtoBuilder.cs が200行上限を超えたため、前例(ElectricToGearDetailDtoBuilder等)に倣いポンプ・採掘機のDTO組み立てをPumpDetailDtoBuilder.cs/MinerDetailDtoBuilder.csへ切り出した。
