+++ moorestech_client/Assets/Scripts/Client.WebUiHost/Game/Topics/BlockDetail/MinerDetailDtoBuilder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4d75be625915345aaaaa9d0e66216587
