+++ moorestech_client/Assets/Scripts/Client.WebUiHost/Game/Topics/BlockDetail/PumpDetailDtoBuilder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cdeffe2808e19460084dfd46e215ee9e
